Corrupted Emperor
For all the drone fans, i made the corrupted emperor WAY more powerfull than the original one, it has less penetration, but the splash after the death of the drones makes the deal. Only 2 drones can kill an player. it may get laggy if "improoved",so,please dont do that ;-;